# backend/lykke/presentation/api/routers/calendars.py
"""Router for Calendar CRUD operations."""

from typing import Annotated
from uuid import UUID

from fastapi import APIRouter, Depends, Query
from lykke.application.commands.calendar import (
    CreateCalendarHandler,
    DeleteCalendarHandler,
    UpdateCalendarHandler,
)
from lykke.application.queries.calendar import (
    GetCalendarHandler,
    SearchCalendarsHandler,
)
from lykke.domain import value_objects
from lykke.domain.entities import CalendarEntity, UserEntity
from lykke.presentation.api.schemas import (
    CalendarCreateSchema,
    CalendarSchema,
    CalendarUpdateSchema,
)
from lykke.presentation.api.schemas.mappers import map_calendar_to_schema

from .dependencies.commands.calendar import (
    get_create_calendar_handler,
    get_delete_calendar_handler,
    get_update_calendar_handler,
)
from .dependencies.queries.calendar import (
    get_get_calendar_handler,
    get_list_calendars_handler,
)
from .dependencies.user import get_current_user

router = APIRouter()


@router.get("/{uuid}", response_model=CalendarSchema)
async def get_calendar(
    uuid: UUID,
    get_calendar_handler: Annotated[
        GetCalendarHandler, Depends(get_get_calendar_handler)
    ],
) -> CalendarSchema:
    """Get a single calendar by ID."""
    calendar = await get_calendar_handler.run(calendar_id=uuid)
    return map_calendar_to_schema(calendar)


@router.get("/", response_model=value_objects.PagedQueryResponse[CalendarSchema])
async def list_calendars(
    list_calendars_handler: Annotated[
        SearchCalendarsHandler, Depends(get_list_calendars_handler)
    ],
    limit: Annotated[int, Query(ge=1, le=1000)] = 50,
    offset: Annotated[int, Query(ge=0)] = 0,
) -> value_objects.PagedQueryResponse[CalendarSchema]:
    """List calendars with pagination."""
    result = await list_calendars_handler.run(
        search_query=value_objects.CalendarQuery(limit=limit, offset=offset),
    )
    paged_response = result
    # Convert entities to schemas
    calendar_schemas = [map_calendar_to_schema(c) for c in paged_response.items]
    return value_objects.PagedQueryResponse(
        items=calendar_schemas,
        total=paged_response.total,
        limit=paged_response.limit,
        offset=paged_response.offset,
        has_next=paged_response.has_next,
        has_previous=paged_response.has_previous,
    )


@router.post("/", response_model=CalendarSchema)
async def create_calendar(
    calendar_data: CalendarCreateSchema,
    user: Annotated[UserEntity, Depends(get_current_user)],
    create_calendar_handler: Annotated[
        CreateCalendarHandler, Depends(get_create_calendar_handler)
    ],
) -> CalendarSchema:
    """Create a new calendar."""
    # Convert schema to entity (id is optional for create, entity will generate it if None)
    calendar = CalendarEntity(
        user_id=user.id,
        name=calendar_data.name,
        auth_token_id=calendar_data.auth_token_id,
        platform_id=calendar_data.platform_id,
        platform=calendar_data.platform,
        last_sync_at=calendar_data.last_sync_at,
    )
    created = await create_calendar_handler.run(calendar=calendar)
    return map_calendar_to_schema(created)


@router.put("/{uuid}", response_model=CalendarSchema)
async def update_calendar(
    uuid: UUID,
    update_data: CalendarUpdateSchema,
    update_calendar_handler: Annotated[
        UpdateCalendarHandler, Depends(get_update_calendar_handler)
    ],
) -> CalendarSchema:
    """Update a calendar."""
    # Convert schema to update object
    from lykke.domain.value_objects import CalendarUpdateObject

    update_object = CalendarUpdateObject(
        name=update_data.name,
        auth_token_id=update_data.auth_token_id,
        last_sync_at=update_data.last_sync_at,
    )
    updated = await update_calendar_handler.run(
        calendar_id=uuid, update_data=update_object
    )
    return map_calendar_to_schema(updated)


@router.delete("/{uuid}", status_code=200)
async def delete_calendar(
    uuid: UUID,
    delete_calendar_handler: Annotated[
        DeleteCalendarHandler, Depends(get_delete_calendar_handler)
    ],
) -> None:
    """Delete a calendar."""
    await delete_calendar_handler.run(calendar_id=uuid)