# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lykke.day",
    "description": "API for managing calendar events",
    "version": "0.1.0"
  },
  "paths": {
...
    "/calendar-entries/today": {
      "get": {
        "tags": [
          "calendar-entries"
        ],
        "summary": "Today",
        "operationId": "today_calendar_entries_today_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CalendarEntrySchema"
                  },
                  "type": "array",
                  "title": "Response Today Calendar Entries Today Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
    "/calendar-entries/today/": {
      "get": {
        "tags": [
          "calendar-entries"
        ],
        "summary": "Today",
        "operationId": "today_calendar_entries_today__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CalendarEntrySchema"
                  },
                  "type": "array",
                  "title": "Response Today Calendar Entries Today  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
...
    "/calendars/{uuid}": {
      "get": {
        "tags": [
          "calendars"
        ],
        "summary": "Get Calendar",
        "description": "Get a single calendar by ID.",
        "operationId": "get_calendar_calendars__uuid__get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "calendars"
        ],
        "summary": "Update Calendar",
        "description": "Update a calendar.",
        "operationId": "update_calendar_calendars__uuid__put",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "calendars"
        ],
        "summary": "Delete Calendar",
        "description": "Delete a calendar.",
        "operationId": "delete_calendar_calendars__uuid__delete",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/calendars/": {
      "get": {
        "tags": [
          "calendars"
        ],
        "summary": "List Calendars",
        "description": "List calendars with pagination.",
        "operationId": "list_calendars_calendars__get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedQueryResponse_CalendarSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "calendars"
        ],
        "summary": "Create Calendar",
        "description": "Create a new calendar.",
        "operationId": "create_calendar_calendars__post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarCreateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActionSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/ActionType"
          },
          "data": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Data"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "type",
          "created_at"
        ],
        "title": "ActionSchema",
        "description": "API schema for Action value object."
      },
      "ActionType": {
        "type": "string",
        "enum": [
          "COMPLETE",
          "DELETE",
          "EDIT",
          "NOTIFY",
          "PAUSE",
          "PUNT",
          "RESUME",
          "START",
          "VIEW"
        ],
        "title": "ActionType"
      },
...
      "CalendarCreateSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "auth_token_id": {
            "type": "string",
            "format": "uuid",
            "title": "Auth Token Id"
          },
          "platform_id": {
            "type": "string",
            "title": "Platform Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "auth_token_id",
          "platform_id",
          "platform"
        ],
        "title": "CalendarCreateSchema",
        "description": "API schema for creating a Calendar entity."
      },
      "CalendarEntrySchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "calendar_id": {
            "type": "string",
            "format": "uuid",
            "title": "Calendar Id"
          },
          "platform_id": {
            "type": "string",
            "title": "Platform Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "frequency": {
            "$ref": "#/components/schemas/TaskFrequency"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionSchema"
            },
            "type": "array",
            "title": "Actions"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "calendar_id",
          "platform_id",
          "platform",
          "status",
          "starts_at",
          "frequency",
          "created_at",
          "updated_at",
          "date"
        ],
        "title": "CalendarEntrySchema",
        "description": "API schema for CalendarEntry entity."
      },
      "CalendarSchema": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "auth_token_id": {
            "type": "string",
            "format": "uuid",
            "title": "Auth Token Id"
          },
          "platform_id": {
            "type": "string",
            "title": "Platform Id"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "auth_token_id",
          "platform_id",
          "platform",
          "user_id"
        ],
        "title": "CalendarSchema",
        "description": "API schema for Calendar entity."
      },
      "CalendarUpdateSchema": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "auth_token_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Token Id"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          }
        },
        "type": "object",
        "title": "CalendarUpdateSchema",
        "description": "API schema for Calendar update requests."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PagedQueryResponse_CalendarSchema_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CalendarSchema"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_next",
          "has_previous"
        ],
        "title": "PagedQueryResponse"
      },
...
      "TaskFrequency": {
        "type": "string",
        "enum": [
          "DAILY",
          "CUSTOM_WEEKLY",
          "WEEKLY",
          "ONCE",
          "YEARLY",
          "MONTHLY",
          "BI_WEEKLY",
          "WORK_DAYS",
          "WEEKENDS"
        ],
        "title": "TaskFrequency"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lykke_auth"
      }
    }
  }
}